pyproject.toml
[project]
name = "jarvis"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "sccuhha" }
]
requires-python = ">=3.12"
dependencies = [
    "ollama>=0.4.8",
    "pyyaml>=6.0.2",
]

[project.scripts]
jarvis = "jarvis.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
